Fix CloudFlare Bypass
Replace hook version with a custom "handler" that can receive a session.
