+++ medusa/session/hooks.py
@@ -4,12 +4,7 @@ from __future__ import unicode_literals
 
 import logging
 
-import cfscrape
-
 from medusa.logger.adapters.style import BraceAdapter
-
-import requests
-from requests.utils import dict_from_cookiejar
 
 from six import text_type
 
@@ -56,76 +51,3 @@ def log_url(response, **kwargs):
         else:
             log.warning('Failed to decode post data with {codecs}',
                         {'codecs': codecs})
-
-
-def cloudflare(resp, **kwargs):
-    """
-    Bypass CloudFlare's anti-bot protection.
-
-    A response hook that retries a request after bypassing CloudFlare anti-bot
-    protection.  Use the sessioned hook factory to attach the session to the
-    response to persist CloudFlare authentication at the session level.
-    """
-    if is_cloudflare_challenge(resp):
-
-        log.debug(u'CloudFlare protection detected, trying to bypass it')
-
-        # Get the session used or create a new one
-        session = getattr(resp, 'session', requests.Session())
-
-        # Get the original request
-        original_request = resp.request
-
-        # Avoid recursion by removing the hook from the original request
-        original_request.hooks['response'].remove(cloudflare)
-
-        # Get the CloudFlare tokens and original user-agent
-        tokens, user_agent = cfscrape.get_tokens(original_request.url)
-
-        # Add CloudFlare tokens to the session cookies
-        session.cookies.update(tokens)
-        # Add CloudFlare Tokens to the original request
-        original_cookies = dict_from_cookiejar(original_request._cookies)
-        original_cookies.update(tokens)
-        original_request.prepare_cookies(original_cookies)
-
-        # The same User-Agent must be used for the retry
-        # Update the session with the CloudFlare User-Agent
-        session.headers['User-Agent'] = user_agent
-        # Update the original request with the CloudFlare User-Agent
-        original_request.headers['User-Agent'] = user_agent
-
-        # Resend the request
-        cf_resp = session.send(
-            original_request,
-            allow_redirects=True,
-            **kwargs
-        )
-
-        if cf_resp.ok:
-            log.debug('CloudFlare successfully bypassed.')
-        return cf_resp
-    else:
-        return resp
-
-
-def sessioned(session):
-    """Hooks factory to add a session to a response."""
-    def sessioned_response_hook(response, *args, **kwargs):
-        """Return a sessioned response."""
-        response.session = session
-        return response
-    return sessioned_response_hook
-
-
-def is_cloudflare_challenge(resp):
-    """Check if the response is a Cloudflare challange.
-
-    Source: goo.gl/v8FvnD
-    """
-    return (
-        resp.status_code == 503
-        and resp.headers.get('Server', '').startswith('cloudflare')
-        and b'jschl_vc' in resp.content
-        and b'jschl_answer' in resp.content
-    )
